Added ability to write and read from flash
The saved configuration should be stored in flash. This
patch includes basic functionality for this. UART interface
is updated with functions to write and read to/from flash.
A flash portion of 4KiB will be available to store persistent
memory and will be located at 0x8007000 together with a
marker/flag at 0x8007C00. The marker will be there to discern
whether custom or default configuration should be loaded at
start up.